tests: Add the tests for the fedimg.utils
